Create component folders and style using vanilla-extract
